fix: conditionally render tabs when FEATURE_BROWSE_AND_REQUEST is enabled
